Shrink the well-known machinery to the one type that reads it
Collapsing the handles onto the prelude scope left a closed set of six
names behind, and only `Array` was ever consulted. `wellKnownPromise`,
`wellKnownIterable`, `wellKnownAsyncIterable`, `wellKnownGenerator` and
`wellKnownAsyncGenerator` had no reader at all: they stayed alive because
`wellKnownNames` listed them and a test walked that list. `wellKnownType`
had one caller, on the line below its own definition. The file read as
though five rules existed that do not.

`resolveArrayClass` now reads `Array` off the prelude scope directly, and
what is left is that plus `arrayElem`, which the rest-parameter,
iteration, spread and indexed-access rules do read. The rules that will
name `Promise` and the iteration protocols arrive in later PRs, and a
constant costs nothing to reintroduce beside its first consumer.

`Context.arrayOf` goes with them. Its one production caller was the
`Array<T>` arm of `resolveTypeAnn`, removed earlier in this PR as
unreachable, which left two tests holding it up. Both build the
`ClassType` directly now, and the negation test reads better for it: it
asserts the run settled a class rather than threading a bool it ignores.

The file is renamed, since `well_known.go` described a set of six and its
subject is one class.

Refs #1497
