+++ internal/solver/array_test.go
@@ -154,19 +154,17 @@ func TestArrayResolvesFromInsideAnOverloadTrial(t *testing.T) {
 	require.NotEmpty(t, c.ctx.arrayClass)
 }
 
-// arrayElem and arrayOf both answer against the class name the run settled on, so
-// both decline when there is nothing to compare against. arrayElem also declines an
-// instance of that class carrying the wrong number of arguments, which is not an
-// array whatever its name says.
+// arrayElem answers against the class name the run settled on, so it declines
+// when there is nothing to compare against. It also declines an instance of that
+// class carrying the wrong number of arguments, which is not an array whatever
+// its name says.
 func TestArrayHelpersDeclineWhatIsNotAnArray(t *testing.T) {
 	t.Parallel()
 
 	t.Run("NoArrayResolved", func(t *testing.T) {
 		t.Parallel()
 		c := &Context{}
-		_, ok := c.arrayOf(num())
-		require.False(t, ok)
-		_, ok = c.arrayElem(&soltype.ClassType{Name: "Array", TypeArgs: []soltype.Type{num()}})
+		_, ok := c.arrayElem(&soltype.ClassType{Name: "Array", TypeArgs: []soltype.Type{num()}})
 		require.False(t, ok)
 	})
 	t.Run("AnotherClassOfTheSameShape", func(t *testing.T) {
